+++ __init__.py
@@ -16,23 +16,12 @@
 #
 # ##### END GPL LICENSE BLOCK #####
 
-
-if "bpy" in locals():
-    import importlib
-    importlib.reload(gob)
-    importlib.reload(gob_import)
-    importlib.reload(preferences)
-    importlib.reload(addon_updater)
-else:
-    from . import gob
-    from . import gob_import
-    from . import preferences
-    from . import addon_updater
+   
 
 import bpy
 import os
 import bpy.utils.previews
-
+from . import gob_import, paths, gob_export, preferences, addon_updater
 
 bl_info = {
     "name": "GoB",
@@ -47,11 +36,11 @@ bl_info = {
 
 
 classes = (
-    gob.GoB_OT_import,
-    gob.GoB_OT_export,
-    gob.GoB_OT_export_button,
-    gob.GoB_OT_GoZ_Installer,
-    gob.GOB_OT_Popup,
+    gob_import.GoB_OT_import,
+    gob_export.GoB_OT_export,
+    gob_import.GoB_OT_export_button,
+    gob_import.GOB_OT_Popup,
+    paths.GoB_OT_GoZ_Installer,
     preferences.GoB_Preferences,
     addon_updater.AU_OT_SearchUpdates,
     )
@@ -66,19 +55,19 @@ def register():
     icons.load("GOZ_SEND", os.path.join(icons_dir, "goz_send.png"), 'IMAGE')
     icons.load("GOZ_SYNC_ENABLED", os.path.join(icons_dir, "goz_sync_enabled.png"), 'IMAGE')
     icons.load("GOZ_SYNC_DISABLED", os.path.join(icons_dir, "goz_sync_disabled.png"), 'IMAGE')
-    gob.preview_collections["main"] = icons 
-    bpy.types.TOPBAR_HT_upper_bar.prepend(gob.draw_goz_buttons)
+    gob_import.preview_collections["main"] = icons 
+    bpy.types.TOPBAR_HT_upper_bar.prepend(gob_import.draw_goz_buttons)
 
 
 def unregister():
 
-    for preferences.custom_icons in gob.preview_collections.values():
+    for preferences.custom_icons in gob_import.preview_collections.values():
         bpy.utils.previews.remove(icons)
-    gob.preview_collections.clear()
+    gob_import.preview_collections.clear()
 
-    bpy.types.TOPBAR_HT_upper_bar.remove(gob.draw_goz_buttons)
+    bpy.types.TOPBAR_HT_upper_bar.remove(gob_import.draw_goz_buttons)
 
     [bpy.utils.unregister_class(c) for c in classes]
 
-    if bpy.app.timers.is_registered(gob.run_import_periodically):
-        bpy.app.timers.unregister(gob.run_import_periodically)
+    if bpy.app.timers.is_registered(gob_import.run_import_periodically):
+        bpy.app.timers.unregister(gob_import.run_import_periodically)
