+++ utils.py
@@ -21,6 +21,7 @@ import numpy
 import random
 import addon_utils
 
+
 def prefs():
     user_preferences = bpy.context.preferences
     return user_preferences.addons[__package__].preferences
